requirements.txt
asgiref==3.8.1
autopep8==2.3.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
